pyproject.toml
[project]
name = "parse-schedule"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bpemb>=0.3.6",
    "google-api-python-client>=2.167.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1",
    "matplotlib>=3.10.1",
    "numpy<2",
    "pandas>=2.2.3",
    "plotly>=6.0.1",
    "python-docx>=1.1.2",
    "scikit-learn>=1.6.1",
    "wordcloud>=1.9.4",
]
